# Optimize WriteTrench performance on Linux