--- a/backend.go
+++ b/backend.go
@@ -394,6 +394,38 @@ func (b *Backend) attachmentReference(name, checksum string) string {
 	return fmt.Sprintf("refs/attachments/%s", h)
 }
 
+func (b *Backend) ListAttachments() ([]Attachment, error) {
+	refs, err := b.r.References()
+	if err != nil {
+		return nil, err
+	}
+	
+	var attachments []Attachment
+	err = refs.ForEach(func(ref *plumbing.Reference) error {
+		if !strings.HasPrefix(ref.Name().String(), "refs/attachments/") {
+			return nil
+		}
+		
+		enc := base64.URLEncoding.WithPadding(base64.NoPadding)
+		encoded := strings.TrimPrefix(ref.Name().String(), "refs/attachments/")
+		decoded, err := enc.DecodeString(encoded)
+		if err != nil {
+			return nil
+		}
+		
+		parts := strings.SplitN(string(decoded), "/", 2)
+		if len(parts) == 2 {
+			attachments = append(attachments, Attachment{
+				Name:     parts[0],
+				Checksum: parts[1],
+			})
+		}
+		return nil
+	})
+	
+	return attachments, err
+}
+
 func (b *Backend) addBlob(data []byte) (plumbing.Hash, error) {
 	obj := b.r.Storer.NewEncodedObject()
 	obj.SetType(plumbing.BlobObject)