- Reduced sync time from ~9 seconds to ~2.3 seconds on Linux servers
      - Optimized git object creation and reference lookups
      - No changes when data hasn't been modified = minimal I/O

      ## Problem
      The WriteTrench function was taking 7-9 seconds on Linux (enki server) vs <1 second on macOS. Investigation revealed:
      1. Linux performs synchronous fsync operations on each git object write
      2. Individual attachment reference lookups were 100x slower on Linux (~1.6ms vs ~15μs)

      ## Solution
      Two key optimizations:

      ### 1. Smart blob creation
      - Check if objects already exist before writing using HasEncodedObject()
      - Avoids unnecessary disk I/O and fsync operations
      - Particularly important on Linux where each write triggers a synchronous fsync

      ### 2. Cache attachment references
      - Pre-load all attachment refs into a map at the start of WriteTrench
      - Eliminates ~867 individual reference lookups per sync
      - Reduces attachment processing time from 1.4s to <10ms

      ## Performance Results
      **Before**: 9+ seconds total sync time
      **After**: 2.3 seconds total sync time

      When no data has changed, WriteTrench now completes in ~130ms vs 7+ seconds previously.

      ## Test Plan
      - [x] Tested on production server (enki.agathe.gr) with 1204 surveys
      - [x] Verified no data corruption or changes when syncing unchanged data
      - [x] Confirmed performance improvement persists across multiple syncs
      - [ ] Test with actual changes to ensure modified surveys are properly written